re added files by upload
i re added the files with noting commit changes to both the files. Style sheet was changed added gradient color background adjusting images to match the alignments, font was resized in the reference section due to words going onto the next line which was out of alignment, which were all adjusted  in the style sheet. The changes made in the index. html file were spelling corrections and grammar.